emma/requirements/base.txt
Django==1.8
uwsgi
psycopg2
pytz
python3-memcached
django-model-utils
django-extensions
django-braces

Pillow
djangorestframework
pysolr
django-haystack
django-geoip
